+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Check for double scrape of the same subscription in the rate limit collector.
+
 ## [2.0.0] - 2020-10-06
 
 ### Changed